LUI-86 Show findpatient nav link only with privilege
"Find Patient" navigation bar/link shows for users that do not have the "View
Patients" privilege. When clicking on it the user just gets redirected to a
page telling him he does not have the privilege to access the page.

* Wrap "Find/Create Patient" navigation bar/link with hasPrivilege
"View Patients" to prevent showing "Find Patient" to a user that doesnt have
the privilege.